refactor: remove src directory and update config paths
