+++ app/api/auth/[...all]/route.ts
@@ -1,4 +1,4 @@
 import { toNextJsHandler } from 'better-auth/next-js';
-import { auth } from '@/features/auth/lib/auth';
+import { auth } from '@/lib/auth';
 
 export const { POST, GET } = toNextJsHandler(auth);
