+++ drizzle.config.ts
@@ -3,7 +3,7 @@ import { defineConfig } from 'drizzle-kit';
 
 export default defineConfig({
   out: './drizzle',
-  schema: './src/db/schema',
+  schema: './db/schema',
   dialect: 'postgresql',
   dbCredentials: {
     url: process.env.DATABASE_URL!
