+++ features/auth/types/index.ts
@@ -1,4 +1,4 @@
-import type { auth } from '../lib/auth';
+import type { auth } from '@/lib/auth';
 
 export type Session = typeof auth.$Infer.Session;
 export type User = typeof auth.$Infer.Session.user;
