+++ lib/auth/client.ts
@@ -1,6 +1,6 @@
 import { adminClient, inferAdditionalFields } from 'better-auth/client/plugins';
 import { createAuthClient } from 'better-auth/react';
-import { type auth } from './auth';
+import { type auth } from '.';
 import { ac, roles } from './permissions';
 
 export const authClient = createAuthClient({
