+++ lib/auth/session.ts
@@ -1,6 +1,6 @@
 import { headers } from 'next/headers';
 import { cache } from 'react';
-import { auth } from './auth';
+import { auth } from '.';
 
 export const getServerSession = cache(async () => {
   return await auth.api.getSession({ headers: await headers() });
